Improve searching of repository configuration
When searching for an scm url from the make mead configuration on PNC to
find a matching repository configuration, PNC will automatically strip
the protocol and '.git' extension and search for the repository
configurations. However on the pnc-cli side, we assume that the
repository configuration received will have the exact same scm url as
what we searched.

This commit relaxes that constraint to consider the repository
configuration received as 'matched' if the configuration's scm url with
the git extension stripped, is the same as the scm url in make-mead with
the git extension stripped.

This commit also renames the git protocol in the make-mead format from
'git+https' to 'https' when 'git+https' is used. This is done since it
seems that more recent versions of git don't support 'git+https' helper
anymore. Given that most make-mead configurations have 'git+https' git
url links in them, this change should help in using the upstream
make-mead configurations without any further modifications, while making
sure PNC will be able to 'git clone' these repositories.