Add support for buildtype for build config
For PNC 1.4, we have added a mandatory field for build configuration
named 'buildType', which specifies whether a build is a maven or npm
build (for now, more options to be added later).

To support this new field, the option '--build_type' has been added to
create a new build configuration (for cli 'create-build-configuration')
and for 'make-mead'.

The default for this option is 'MVN'.

Still on the TODO list: Investigate how this new type affect the Maitai
process and if we need to do adjustments to the bpmconfiguration
creation on make-mead.